Add documentation - rewards in hf19
The reward system is a part of Steem that is changed constantly every hardfork. If you find a post explaining the details be careful because maybe the post is from last year and therefore obsolete. How to deal with that? we need more documentation on the github repository explaining the details of every change. This work consist in a pull request explaining the reward system for authors and curators in HF19.